move notification helper functions out of errors.ts and into notifications.ts
